+++ Tests/New-BicepParameterFile.Tests.ps1
@@ -42,7 +42,7 @@ Describe 'New-BicepParameterFile' {
                 New-BicepParameterFile -Path 'TestDrive:\brokenBicep.bicep' -ErrorAction Stop
             }
             catch {
-                $_.Exception.Message | Should -BeLike '*Error BCP018: Expected the "}" character at this location.'
+                $_.Exception.Message | Should -BeLike '*Error BCP018: Expected the "}" character at this location. `[https://aka.ms/bicep/core-diagnostics#BCP018]'
             }
         }
     }